Move MultiTrie into common separate module
